ref(core): consolidate menuListItem component files and remove chonk naming
Moves all content from menuListItem.chonk.tsx into menuListItem.tsx. Removes
"Chonk" prefixes from all styled components (ChonkInnerWrap, ChonkContentWrap,
ChonkLeadingItems, ChonkLabel, ChonkLabelWrap, ChonkDetails).
